executable: extract method Daemon._terminate()
Some programs run as daemons might be unhappy to be terminated by
SIGTERM (as done by self._process.terminate()). E.g. when running
`perf mem record`, when it is terminated by SIGTERM, it does not
save the profiling data. It needs to be terminated by SIGINT instead.

The method _terminate() can be overriden to custmize the termination
mechanism.