use different images for tests and other jobs
Since change of docker image from python36 to oraclelinux:8
package built worked incorrectly. In particular, setuptools-git-versioning
was unable to compose right package name. It appears that this
bug is combination of two aspects:
- missing git
- version of setuptools-git-versioning including and above 1.8.0
is not working on oraclelinux:8. Probably because of older version
of setuptools which is part of the system in this image.

Reason to transit from python36 to oraclelinux:8 image was
the pytest job. Thus, since now, pytest job will use
oraclelinux:8 image while rest of the jobs will use former
python36.