spirent: introduce method Spirent.determine_src_mac_address()
This method is useful whenever there are more spirent ports that may
use the same source MAC address for their stream blocks. In that case,
the connected switch may infer the MAC address for one of the ports
and then discard all packets coming from another port with the same
source MAC address.