update to python3.8
We are also transfering from installing python dependencies system wide
to usage of virtual environment. With newer versions of pipenv (>2022.8.5)
we are struggling with faulty python detection when using "--system"
argument of pipenv. As we do not want to stuck with one exact version
of pipenv we are forced to use a virtual environment(s) for installation
of python dependencies and usage of our tools and tests.