executable: ensure command is running after Daemon.start()
If an execution of a command is possible but command fails for some
reason, underlying subprocess.Popen() does not throw any exception.
However, we want to have a command always running after calling
of Daemon.start() method. Thus, we are now automatically checking
whether the command is running after the start and thrown an exception
if it is not.