executable: add option to treat SIGTERM as normal termination
Some programs terminate on SIGTERM with -15 [1] (143 [2]) return code
(e.g. ping). As it is non-zero return code, it is treated as a failure
even when these programs actually do not fail. Thus, a new argument
"sigterm_ok" (False by default) is introduced to the Executable
constructor. If this argument is set to True, a command will not
fail in "died with <Signals.SIGTERM: 15>" situation.

[1] If process is terminated by signal, return code is a negative signal
number (https://docs.python.org/3/library/subprocess.html#subprocess.CalledProcessError.returncode)

[2] 143 is returned in the system (128 + signal number)