trex: introduce stateless class
Module implements class for stateless TRex generator.

In this mode, TRex does not store any connection state.
It does not react to current situation on a network.
It is essentially a packet generator similar to Scapy or Spirent.

Generated traffic is configured via TRexStream object.
This object is then loaded into TRex. Generation can be
started via start() method.