ci: utilize code coverage
Report is stored as HTML for user and XML for CI.
GitLab can use XML in Cobertura format to display
which lines of code are covered and which are not.

As described in [1], paths in XML file should be
relative to the project root. Coverage by default
produces absolute paths. GitLab then can't match files
and coverage for each line of code is not displayed.

For this reason we must remove part of absolute path from
generated XML file. This is done using 'sed' tool.
Same thing is done for HTML report as well. Although
it's not necessary, it produces shorter file names
and HTML report looks better.

[1] https://docs.gitlab.com/ee/ci/testing/test_coverage_visualization.html#automatic-class-path-correction