machine_defaults: set timeout for urlopen()
The default timeout of urllib.request.urlopen() is set as
socket._GLOBAL_DEFAULT_TIMEOUT[1]. The value returned by
socket.getdefaulttimeout() is None, which means no timeout[2].

This commit sets timeout to 60s, so pytest fails quickly in case URL is
unreachable instead of waiting forever. Additionally, timeout
exception is raised, so it's easier to troubleshoot the problem.

In case of timeout exception, the traceback looks like this:

  Traceback (most recent call last):
    ...
      File "/usr/local/lib/python3.7/socket.py", line 589, in readinto
        return self._sock.recv_into(b)
    socket.timeout: timed out

[1] https://github.com/python/cpython/blob/3.12/Lib/urllib/request.py#L138
[2] https://docs.python.org/3/library/socket.html#socket.getdefaulttimeout