pytest_plugins: load spirent config file from test directory
Originally, the spirent configuration file was expected to exist in the
fixtures file directory. As we are extrating this fixture into a separate
pytest plugin, this no longer makes sense. This commit modifies the
search path to the 'request.node.path.parent' directory, therefore
pytest searches for configuration path in the invoking test module
directory.