trex_packet_crafter: utilize value distribution for l4 ports
Distribution is not utilized for L3 due to limitations
for IPv6 addresses. As IPv6 addresses can contain
two instructions, random distribution would mean first
half of IPv6 would be matched with random second half
of IPv6. This would cause mismatch if IPv6 addresses are
provided as list of exact addresses.