trex_stateful: always set ip.tos header to 0x2
Some NICs (e.g. ConnectX-6) may modify the lowest bit of the TOS header.
Specifically, the ECN bit 0 would be set to 1. This, however, breaks TRex server
as it will filter all packets with this bit set. As a workaround, we explicitly
set the ip.tos tunable parameter to 0x2 (0b10) of the TOS header. This way, the
NIC does not set the extra bit and TRex server can process the packets.