trex_generator: freeing generator returns resources to original order
Network interfaces of trex machine were returned in reverse order.
For example, original order is [A, B, C] and first interface (A)
is requested. When freed, it was appended back to list, resulting in [B, C, A].

Subsequent test case then used different interfaces compared to
previous test case. There is now a need to preserve original
order of interfaces when all requested generators are freed.

For consistency apply same threatment to cores as well.
Daemons are already returned to original order. Upon
request, last deamon in list is selected (pop()) so
appending daemon back to the end of list keeps order.