trex: support two numa nodes and specific cores
Goal of this commit is to support machine with two
NUMA nodes. Configuration file that is automatically
generated is currently hardwired to always use
NUMA node 0 or also known as socket 0.

When using PCI interface that was in NUMA node 1, then
TRex would fail with "mlx5_common: Failed to initialize
global MR share cache." error.

First change is that --trex-generator parameter now
supports optional NUMA node ID. This "binds" interface
to given node.

Generation of configuration file is reworked to use
NUMA node information in mind. This mainly done in
"_setup_cpu_cores()" function.

Second change is ability to specify which cores to use.
Previously only number of cores was supported.
CPU cores are assigned to NUMA nodes, so this is
related to first change as best performace is
possible only when cores are assigned properly.