topology: fix exception when iterating through topology with no generator
When a topology has no generator (such as 'vdev_ring'), the
_generators_by_type() function raises the following exception:

	self = <lbr_testsuite.topology.topology.Topology object at 0x7f8312e4e410>, gen_type = <class 'lbr_testsuite.topology.generator.Generator'>
	    def _generators_by_type(self, gen_type):
	>       return [g for g in self._generators if isinstance(g, gen_type)]
	E       TypeError: 'NoneType' object is not iterable

This error is caused by the '_generators' attribute being set to None
in initialization. This commit fixes the issue by explicitly setting the
'_generators' attribute to an empty tuple when no generators are
provided - empty tuple is iterable and does not raise exceptions.

Fixes: f31c02c1 ("topology: support a tuple of generators in a Topology")
Co-authored-by: Jan Kučera <jan.kucera@cesnet.cz>